Move dataset to nlp/
