+++ mlbench_core/models/pytorch/gnmt/encoder.py
@@ -1,8 +1,8 @@
-import mlbench_core.dataset.translation.pytorch.config as config
 import torch.nn as nn
+from torch.nn.utils.rnn import pack_padded_sequence, pad_packed_sequence
+
+import mlbench_core.dataset.nlp.pytorch.translation.wmt14_config as config
 from mlbench_core.models.pytorch.gnmt.utils import init_lstm_
-from torch.nn.utils.rnn import pack_padded_sequence
-from torch.nn.utils.rnn import pad_packed_sequence
 
 
 class ResidualRecurrentEncoder(nn.Module):
